+++ src/db/parity_db.rs
@@ -10,8 +10,7 @@ use super::EthMappingsStore;
 use crate::db::{parity_db_config::ParityDbConfig, DBStatistics, GarbageCollectable};
 use crate::libp2p_bitswap::{BitswapStoreRead, BitswapStoreReadWrite};
 
-use crate::rpc::eth;
-
+use crate::rpc::eth::types::EthHash;
 use anyhow::{anyhow, Context as _};
 use cid::multihash::Code::Blake2b256;
 
@@ -177,15 +176,15 @@ impl SettingsStore for ParityDb {
 }
 
 impl EthMappingsStore for ParityDb {
-    fn read_bin(&self, key: &eth::Hash) -> anyhow::Result<Option<Vec<u8>>> {
+    fn read_bin(&self, key: &EthHash) -> anyhow::Result<Option<Vec<u8>>> {
         self.read_from_column(key.0.as_bytes(), DbColumn::EthMappings)
     }
 
-    fn write_bin(&self, key: &eth::Hash, value: &[u8]) -> anyhow::Result<()> {
+    fn write_bin(&self, key: &EthHash, value: &[u8]) -> anyhow::Result<()> {
         self.write_to_column(key.0.as_bytes(), value, DbColumn::EthMappings)
     }
 
-    fn exists(&self, key: &eth::Hash) -> anyhow::Result<bool> {
+    fn exists(&self, key: &EthHash) -> anyhow::Result<bool> {
         self.db
             .get_size(DbColumn::EthMappings as u8, key.0.as_bytes())
             .map(|size| size.is_some())
@@ -206,7 +205,7 @@ impl EthMappingsStore for ParityDb {
         Ok(cids)
     }
 
-    fn delete(&self, keys: Vec<eth::Hash>) -> anyhow::Result<()> {
+    fn delete(&self, keys: Vec<EthHash>) -> anyhow::Result<()> {
         Ok(self.db.commit_changes(keys.into_iter().map(|key| {
             let bytes = key.0.as_bytes().to_vec();
             (DbColumn::EthMappings as u8, Operation::Dereference(bytes))