+++ src/tool/subcommands/api_cmd.rs
@@ -1277,6 +1277,22 @@ fn eth_tests() -> Vec<RpcTest> {
             )
             .unwrap(),
         ));
+        tests.push(RpcTest::basic(
+            EthNewFilter::request_with_alias(
+                (EthFilterSpec {
+                    from_block: None,
+                    to_block: None,
+                    address: vec![EthAddress::from_str(
+                        "0xff38c072f286e3b20b3954ca9f99c05fbecc64aa",
+                    )
+                    .unwrap()],
+                    topics: EthTopicSpec(vec![]),
+                    block_hash: None,
+                },),
+                use_alias,
+            )
+            .unwrap(),
+        ));
         tests.push(RpcTest::identity(
             EthAddressToFilecoinAddress::request((EthAddress::from_str(
                 "0xff38c072f286e3b20b3954ca9f99c05fbecc64aa",
@@ -1290,7 +1306,7 @@ fn eth_tests() -> Vec<RpcTest> {
 
 fn eth_tests_with_tipset<DB: Blockstore>(store: &Arc<DB>, shared_tipset: &Tipset) -> Vec<RpcTest> {
     let block_cid = shared_tipset.key().cid().unwrap();
-    let block_hash: Hash = block_cid.into();
+    let block_hash: EthHash = block_cid.into();
 
     let mut tests = vec![
         RpcTest::identity(
@@ -1478,7 +1494,7 @@ fn eth_state_tests_with_tipset<DB: Blockstore>(
         }
     }
     tests.push(RpcTest::identity(
-        EthGetMessageCidByTransactionHash::request((Hash::from_str(
+        EthGetMessageCidByTransactionHash::request((EthHash::from_str(
             "0x37690cfec6c1bf4c3b9288c7a5d783e98731e90b0a4c177c2a374c7a9427355f",
         )?,))?,
     ));